URLComponentTests
