+++ ReactiveAPITests/Resources/Resources.swift
@@ -5,6 +5,9 @@ struct Resources {
     static let baseUrl = URL(string: "https://baseurl.com/")!
     static let urlRequest = URLRequest(url: Resources.url)
     static let data = Data(count: 100)
+    static let params: [String : Any?] = ["key": "value",
+                                          "number": 3,
+                                          "nil": nil]
     
     static func httpUrlResponse(code: Int = 200) -> HTTPURLResponse? {
         return HTTPURLResponse(url: Resources.url,
